% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-permute.R
\name{calculateOR}
\alias{calculateOR}
\title{Calculate Odds Ratio}
\usage{
calculateOR(A, G)
}
\arguments{
\item{A}{An adjacency matrix.}

\item{G}{An adjacency matrix.}
}
\description{
This function calculates an odds ratio for the
 edge overlap between two non-random graphs.
}
\details{
Note that this function calculates overlap for the
 lower-left triangle of the input matrices.
}
