% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-graflex.R
\name{getK}
\alias{getK}
\title{Build Knowledge Database for Human Genes}
\usage{
getK(keys, columns = "GO", keytype = "ENSEMBL", minK = 10)
}
\arguments{
\item{keys}{A character vector. The IDs for the genes to map.
Argument passed to \code{AnnotationDbi::select}.}

\item{columns}{A string. The concepts to return (e.g., GO).
Argument passed to \code{AnnotationDbi::select}.}

\item{keytype}{A string. The type of IDs being mapped.
Argument passed to \code{AnnotationDbi::select}.}

\item{minK}{An integer. The minimum number of nodes to which
a concept maps to be included in final table.
Skip with \code{minK = 0}.}
}
\value{
A knowledge database where each row is a graph node
 and each column is a concept.
}
\description{
This helper function wraps the few steps needed
 to make a "knowledge database" used by \code{\link{graflex}}.
}
\details{
A knowledge database contains a row (g) for each node in the graph,
 and a column (c) for each concept tested for enrichment.
 \code{K_{gc} = 1} whenever a node associates with a concept.
}
