% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-backend.R
\name{getOR}
\alias{getOR}
\title{Calculate Odds Ratio}
\usage{
getOR(A, G)
}
\arguments{
\item{A}{A vector or adjacency matrix.}

\item{G}{A vector or adjacency matrix.}
}
\value{
A \code{data.frame} of results.
}
\description{
This function calculates the overlap between
 two vectors or two adjacency matrices.
 It returns the OR as well as other metrics.
}
