% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-randFuncs.R
\name{randGraph}
\alias{randGraph}
\title{Make Random Graph}
\usage{
randGraph(size)
}
\arguments{
\item{size}{An integer. The size of the graph.}
}
\value{
An adjacency matrix.
}
\description{
This function makes a random graph.
}
