% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-randFuncs.R
\name{randK}
\alias{randK}
\title{Make Random Knowledge Database}
\usage{
randK(nnodes, nconcepts, fill = 0.5 * nnodes)
}
\arguments{
\item{nnodes}{The number of nodes (rows).}

\item{nconcepts}{The number of concepts (columns).}

\item{fill}{An integer. The number of nodes to connect.}
}
\value{
A knowledge database where each row is a graph node
 and each column is a concept.
}
\description{
This function builds a knowledge database with nodes
 as rows and concepts as columns.
}
