% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-randFuncs.R
\name{randPropd}
\alias{randPropd}
\title{Make Random Propd Graph}
\usage{
randPropd(size)
}
\arguments{
\item{size}{An integer. The size of the graph.}
}
\value{
An adjacency matrix.
}
\description{
This function builds a graph based on a \code{propr::propd}
 analysis of a random Poisson matrix.
}
